fix: modifyBundlerChainHook
