+++ packages/builder/builder-rspack-provider/src/types/plugin.ts
@@ -1,7 +1,6 @@
 import type {
   DefaultBuilderPluginAPI,
   BuilderPlugin as BaseBuilderPlugin,
-  ModifyBundlerChainFn,
 } from '@modern-js/builder-shared';
 import type { BuilderConfig, NormalizedConfig } from './config';
 import type { ModifyRspackConfigFn } from './hooks';
@@ -15,7 +14,6 @@ export interface BuilderPluginAPI
     Compiler
   > {
   modifyRspackConfig: (fn: ModifyRspackConfigFn) => void;
-  modifyBundlerChain: (fn: ModifyBundlerChainFn) => void;
 }
 
 export type BuilderPlugin = BaseBuilderPlugin<BuilderPluginAPI>;
