+++ packages/builder/builder-shared/src/types/bundlerConfig.ts
@@ -1,7 +1,5 @@
 import { Configuration } from 'webpack';
 import type Config from '../../compiled/webpack-5-chain';
-import { NodeEnv } from './utils';
-import { BuilderTarget } from './builder';
 
 /** The intersection of webpack and rspack */
 export type BundlerConfig = {
@@ -28,17 +26,3 @@ export interface BundlerChain
   extends Pick<Config, 'devtool' | 'target' | 'name' | 'merge'> {
   toConfig: () => BundlerConfig;
 }
-
-export type ModifyBundlerChainUtils = {
-  env: NodeEnv;
-  isProd: boolean;
-  target: BuilderTarget;
-  isServer: boolean;
-  isWebWorker: boolean;
-};
-
-/** The intersection of webpack-chain and rspack-chain */
-export type ModifyBundlerChainFn = (
-  chain: BundlerChain,
-  utils: ModifyBundlerChainUtils,
-) => Promise<void> | void;
