+++ packages/builder/builder-shared/src/types/hooks.ts
@@ -1,4 +1,7 @@
 import type { Stats, MultiStats } from './stats';
+import { NodeEnv } from './utils';
+import { BuilderTarget } from './builder';
+import { BundlerChain } from './bundlerConfig';
 
 export type OnBeforeBuildFn<BundlerConfig = unknown> = (params: {
   bundlerConfigs?: BundlerConfig[];
@@ -31,3 +34,22 @@ export type OnExitFn = () => void;
 export type ModifyBuilderConfigFn<BuilderConfig> = (
   config: BuilderConfig,
 ) => Promise<BuilderConfig | void> | BuilderConfig | void;
+
+export type ModifyBundlerChainUtils = {
+  env: NodeEnv;
+  isProd: boolean;
+  target: BuilderTarget;
+  isServer: boolean;
+  isWebWorker: boolean;
+};
+
+/** The intersection of webpack-chain and rspack-chain */
+export type ModifyBundlerChainFn = (
+  chain: BundlerChain,
+  utils: ModifyBundlerChainUtils,
+) => Promise<void> | void;
+
+export type CreateAsyncHook<Callback extends (...args: any[]) => any> = {
+  tap: (cb: Callback) => void;
+  call: (...args: Parameters<Callback>) => Promise<Parameters<Callback>>;
+};
